+++ gcl/h/bsd.h
@@ -44,8 +44,8 @@ extern char etext;
 		real_maxpage = MAXPAGE ; } while(0)
      
 #define ROUND_UP_SBRK(x)  \
-       do {int i; \
-	     if (i = ((int)x & (PAGESIZE - 1))) \
+       do {long i; \
+	     if (i = ((long)x & (PAGESIZE - 1))) \
 	       x=sbrk(PAGESIZE - i); } while(0);
 
 #define FIX_RANDOM_SBRK \
